added multi-cluster distributed

+++ argocd/iac/terraform/examples/eks/multi-cluster-distributed/destroy.sh
@@ -1,7 +1,18 @@
 #!/bin/bash
 
+
+if [[ $# -eq 0 ]] ; then
+    echo "No arguments supplied"
+    echo "Usage: destroy.sh <environment>"
+    echo "Example: destroy.sh dev"
+    exit 1
+fi
+env=$1
+echo "Destroying $env ..."
+
 set -x
 
+terraform workspace select $env
 terraform destroy -target="module.gitops_bridge_bootstrap" -auto-approve
 terraform destroy -target="module.eks_blueprints_addons" -auto-approve
 terraform destroy -target="module.eks" -auto-approve
